USCS Extraction of Germline Mutations
This program runs over a file of Mutations, and extracts data from the dbSNP track to check whether each mutation is Somatic or Germline.

It also iterates over the nucleotide positions before and after the mutation's nucleotide position and compares whether the mutation's ref and alt alleles are identical to the ref and alt of either nucleotide position before or after respectively. If so these mutations are also flagged as Germline mutations.